// src/app/models/leave/leaveAppLineContract.interface.ts
export class LeaveAppLineContract {
    // AbsenceCode:string;
    // AbsenceCodeDescription:string;
    // BalanceDays:Number;
    // hours:Number;
    // IsDeleted:boolean;
    // LeaveDays:Number;
    // Remarks:String;
    // ValidFrom:Date;
    // ValidTo:Date;

    TableNumber:any;
    AbsenceCode: any;
    AbsenceCodeDescription: any;
    ActualEndDate: any;
    ActualStartDate: any;
    Balance: any;
    EndDate: any;
    Hours: any;
    LeaveApplicationCode: any;
    LeaveApplied: any;
    Remarks: any;
    StartDate: any;
    DataArea:any;
    ERPRecID:any;

    IsDeleted:boolean;
}